+++ routes/index.js
@@ -1,0 +1,12 @@
+import { Router } from 'express'
+import booksRoutes from './books.js'
+import usersRoutes from './users.js'
+
+const router = Router()
+
+router.get('/', (req, res) => res.send('This is the api root!'))
+
+router.use('/', usersRoutes)
+router.use('/', booksRoutes)
+
+export default router
